// src/services/pokemon/type.d.ts
type PokemonResponse = {
    id: number
    name: string
    base_experience: number
    height: number
    is_default: boolean
    order: number
    weight: number
    abilities: Ability[]
    forms: Form[]
    game_indices: Index[]
    held_items: HeldItem[]
    location_area_encounters: string
    moves: Mfe[]
    species: Species
    cries: Cries
    stats: Stat[]
    types: Type[]
    past_types: PastType[]
    sprites: Sprites

    count: number
    next: string
    previous: any
    results: Pokemons[]
  }

  type Pokemons = {
    name: string
    url: string
  }
  
  type Ability = {
    results: Poke[]
    is_hidden: boolean
    slot: number
    ability: Ability2
  }
  
  type Ability2 = {
    name: string
    url: string
  }
  
  type Form = {
    name: string
    url: string
  }
  
  type Index = {
    game_index: number
    version: Version
  }
  
  type Version = {
    name: string
    url: string
  }
  
  type HeldItem = {
    item: Item
    version_details: VersionDetail[]
  }
  
  export type Item = {
    name: string
    url: string
  }
  
  export type VersionDetail = {
    rarity: number
    version: Version2
  }
  
  export type Version2 = {
    name: string
    url: string
  }
  
  export type Mfe = {
    move: Move
    version_group_details: VersionGroupDetail[]
  }
  
  export type Move = {
    name: string
    url: string
  }
  
  export type VersionGroupDetail = {
    level_learned_at: number
    version_group: VersionGroup
    move_learn_method: MoveLearnMethod
  }
  
  export type VersionGroup = {
    name: string
    url: string
  }
  
  export type MoveLearnMethod = {
    name: string
    url: string
  }
  
  export type Species = {
    name: string
    url: string
  }
  
  export type Sprites = {
    back_default: string
    back_female: any
    back_shiny: string
    back_shiny_female: any
    front_default: string
    front_female: any
    front_shiny: string
    front_shiny_female: any
    other: Other
    versions: Versions
  }
  
  type Other = {
    dream_world: DreamWorld
    home: Home
    "official-artwork": OfficialArtwork
    showdown: Showdown
  }
// ...
  type Versions = {
    "generation-i": GenerationI
    "generation-ii": GenerationIi
    "generation-iii": GenerationIii
    "generation-iv": GenerationIv
    "generation-v": GenerationV
    "generation-vi": GenerationVi
    "generation-vii": GenerationVii
    "generation-viii": GenerationViii
  }
// ...
  type Cries = {
    latest: string
    legacy: string
  }
  
  type Stat = {
    base_stat: number
    effort: number
    stat: Stat2
  }
  
  type Stat2 = {
    name: string
    url: string
  }
  
  type Type = {
    slot: number
    type: Type2
  }
  
  type Type2 = {
    name: string
    url: string
  }
  
  type PastType = {
    generation: Generation
    types: Type3[]
  }
  
  type Generation = {
    name: string
    url: string
  }
  
  type Type3 = {
    slot: number
    type: Type4
  }